Writing Flow: Move selected block focus to BlockListBlock
